# public/swagger.json
{
	"components": {
...
		"schemas": {
...
			"ShipmentModel": {
				"properties": {
					"shipmentid": {
						"type": "number",
						"format": "double"
					},
					"sequencenumber": {
						"type": "number",
						"format": "double"
					},
					"date": {
						"type": "string",
						"format": "date-time"
					},
					"modality": {
						"type": "string"
					},
					"distance": {
						"type": "number",
						"format": "double"
					},
					"intervaldistance": {
						"type": "string"
					},
					"emissionprocessing": {
						"type": "number",
						"format": "double"
					},
					"fefrom": {
						"type": "string"
					},
					"feto": {
						"type": "string"
					},
					"routeid": {
						"type": "string"
					},
					"fromclient": {
						"type": "string"
					},
					"toclient": {
						"type": "string"
					},
					"containertype": {
						"type": "string"
					},
					"intervalweight": {
						"type": "string"
					},
					"transporttype": {
						"type": "string"
					},
					"shipmentweight": {
						"type": "number",
						"format": "double"
					},
					"teu": {
						"type": "number",
						"format": "double"
					},
					"co2pershipment": {
						"type": "number",
						"format": "double"
					},
					"Container": {
						"$ref": "#/components/schemas/ContainerModel"
					},
					"CO2": {
						"$ref": "#/components/schemas/CO2Model"
					},
					"NOX": {
						"$ref": "#/components/schemas/NOXModel"
					}
				},
				"required": [
					"shipmentid",
					"sequencenumber",
					"date",
					"modality",
					"distance",
					"intervaldistance",
					"emissionprocessing",
					"fefrom",
					"feto",
					"routeid",
					"fromclient",
					"toclient",
					"containertype",
					"intervalweight",
					"transporttype",
					"shipmentweight",
					"teu",
					"co2pershipment",
					"Container",
					"CO2",
					"NOX"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ContainerModel": {
				"properties": {
					"containertype": {
						"type": "string"
					},
					"lengthcontainer": {
						"type": "number",
						"format": "double"
					},
					"widthcontainer": {
						"type": "number",
						"format": "double"
					},
					"heightcontainer": {
						"type": "number",
						"format": "double"
					},
					"Shipment": {
						"items": {
							"$ref": "#/components/schemas/ShipmentModel"
						},
						"type": "array"
					}
				},
				"required": [
					"containertype",
					"lengthcontainer",
					"widthcontainer",
					"heightcontainer",
					"Shipment"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CO2Model": {
				"properties": {
					"emissionsperkm": {
						"type": "number",
						"format": "double"
					},
					"transporttype": {
						"type": "string"
					},
					"fe": {
						"type": "string"
					},
					"intervalweight": {
						"type": "string"
					},
					"teu": {
						"type": "number",
						"format": "double"
					},
					"Shipment": {
						"items": {
							"$ref": "#/components/schemas/ShipmentModel"
						},
						"type": "array"
					}
				},
				"required": [
					"emissionsperkm",
					"transporttype",
					"fe",
					"intervalweight",
					"teu",
					"Shipment"
				],
				"type": "object",
				"additionalProperties": false
			},
			"NOXModel": {
				"properties": {
					"emissionsperkm": {
						"type": "number",
						"format": "double"
					},
					"transporttype": {
						"type": "string"
					},
					"intervalweight": {
						"type": "string"
					},
					"teu": {
						"type": "number",
						"format": "double"
					},
					"Shipment": {
						"items": {
							"$ref": "#/components/schemas/ShipmentModel"
						},
						"type": "array"
					}
				},
				"required": [
					"emissionsperkm",
					"transporttype",
					"intervalweight",
					"teu",
					"Shipment"
				],
				"type": "object",
				"additionalProperties": false
			},
			"RoutesModel": {
				"properties": {
					"routeid": {
						"type": "string"
					},
					"modality": {
						"type": "string"
					},
					"fueltype": {
						"type": "string"
					},
					"TransportType": {
						"type": "string"
					}
				},
				"required": [
					"routeid",
					"modality",
					"fueltype",
					"TransportType"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Shipment": {
				"properties": {
					"co2pershipment": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"teu": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"shipmentweight": {
						"type": "number",
						"format": "double"
					},
					"transporttype": {
						"type": "string"
					},
					"intervalweight": {
						"type": "string",
						"nullable": true
					},
					"containertype": {
						"type": "string"
					},
					"toclient": {
						"type": "string"
					},
					"fromclient": {
						"type": "string"
					},
					"routeid": {
						"type": "string"
					},
					"feto": {
						"type": "string"
					},
					"fefrom": {
						"type": "string"
					},
					"emissionprocessing": {
						"type": "number",
						"format": "double"
					},
					"intervaldistance": {
						"type": "string"
					},
					"distance": {
						"type": "number",
						"format": "double"
					},
					"modality": {
						"type": "string"
					},
					"date": {
						"type": "string",
						"format": "date-time"
					},
					"sequencenumber": {
						"type": "number",
						"format": "double"
					},
					"shipmentid": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"co2pershipment",
					"teu",
					"shipmentweight",
					"transporttype",
					"intervalweight",
					"containertype",
					"toclient",
					"fromclient",
					"routeid",
					"feto",
					"fefrom",
					"emissionprocessing",
					"intervaldistance",
					"distance",
					"modality",
					"date",
					"sequencenumber",
					"shipmentid"
				],
				"type": "object",
				"description": "Model Shipment"
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "api",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
...
		"/nox": {
			"get": {
				"operationId": "GetNoxs",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/NOXModel"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Nox"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "offset",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"post": {
				"operationId": "CreateNox",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/NOXModel"
								}
							}
						}
					}
				},
				"tags": [
					"Nox"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/NOXModel"
							}
						}
					}
				}
			},
			"put": {
				"operationId": "UpdateNox",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/NOXModel"
								}
							}
						}
					}
				},
				"tags": [
					"Nox"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/NOXModel"
							}
						}
					}
				}
			}
		},
		"/nox/query": {
			"get": {
				"operationId": "GetNox",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/NOXModel"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Nox"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "intervalweight",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "teu",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "transporttype",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"delete": {
				"operationId": "DeleteNox",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/NOXModel"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Nox"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "intervalweight",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "teu",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "transporttype",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/Route/{routeId?}": {
			"get": {
				"operationId": "GetRoutes",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/RoutesModel"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Route"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "offset",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "path",
						"name": "routeId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/Route": {
			"post": {
				"operationId": "CreateRoute",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/RoutesModel"
								}
							}
						}
					}
				},
				"tags": [
					"Route"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/RoutesModel"
							}
						}
					}
				}
			},
			"put": {
				"operationId": "UpdateRoute",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/RoutesModel"
								}
							}
						}
					}
				},
				"tags": [
					"Route"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/RoutesModel"
							}
						}
					}
				}
			}
		},
		"/Route/{routeId}": {
			"delete": {
				"operationId": "DeleteRoute",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/RoutesModel"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Route"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "routeId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/shipment/{id?}": {
			"get": {
				"operationId": "GetShipments",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/ShipmentModel"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Shipment"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "offset",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/shipment": {
			"post": {
				"operationId": "CreateShipment",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ShipmentModel"
								}
							}
						}
					}
				},
				"tags": [
					"Shipment"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Shipment"
							}
						}
					}
				}
			},
			"put": {
				"operationId": "UpdateShipment",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ShipmentModel"
								}
							}
						}
					}
				},
				"tags": [
					"Shipment"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Shipment"
							}
						}
					}
				}
			}
		},
		"/shipment/{id}": {
			"delete": {
				"operationId": "DeleteShipment",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/ShipmentModel"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Shipment"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}